requirements.txt
asgi-redis==0.14.1
asgiref==0.14.0
asn1crypto==0.24.0
autobahn==0.16.0
certifi==2018.11.29
cffi==1.11.5
channels==0.17.3
cloudinary==1.14.0
colorama==0.3.7
contextlib2==0.5.4
cryptography==2.3.1
daphne==0.14.3
defusedxml==0.5.0
dj-database-url==0.4.1
Django==2.0
django-allauth==0.38.0
django-bootstrap4==0.0.7
django-braces==1.13.0
django-extensions==2.1.3
django-geoposition==0.3.0
django-heroku==0.3.1
django-idmap==1.0.3
django-map-widgets==0.2.1
django-sslserver==0.20
django-vanilla-views==1.0.4
grep==0.3.2
gunicorn==19.9.0
honcho==0.7.1
huey==1.10.2
idna==2.7
IPy==0.83
mock==2.0.0
msgpack-python==0.4.8
oauthlib==2.1.0
otree==2.1.22
otree-boto2-shim==0.3.2
otree-core==0.0.0b1
otree-save-the-change==1.1.3
pbr==1.10.0
Pillow==5.3.0
psycopg2==2.7.5
py==1.4.31
pycparser==2.19
pyOpenSSL==18.0.0
pytest==2.9.2
pytest-django==3.0.0
python-redis-lock==3.2.0
python3-openid==3.1.0
pytz==2017.3
raven==6.9.0
redis==2.10.5
requests==2.11.1
requests-oauthlib==1.0.0
schema==0.6.2
six==1.10.0
termcolor==1.1.0
Twisted==16.2.0
txaio==2.5.1
unicodecsv==0.14.1
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==3.2.1
ws4py==0.3.5
XlsxWriter==0.9.3
zope.interface==4.2.0
